@@ -4,25 +4,25 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.9.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+
+  
 
 dependencies:
   flutter:
     sdk: flutter
-  chopper: ^3.0.6
-  provider: ^4.0.5
+  chopper: ^4.0.0-nullsafety.0
   cupertino_icons: ^0.1.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  chopper_generator: ^3.0.6
-  json_annotation: ^3.0.1
-  json_serializable: ^3.5.0
   swagger_dart_code_generator:
     path: ../
-  build_runner: ^1.10.1
-  
+  chopper_generator: ^4.0.0-nullsafety.0
+  build_runner: ^1.12.2
+  json_annotation: ^4.0.0
+  json_serializable: ^4.0.0
 
 flutter:
   uses-material-design: true
