@@ -8,17 +8,17 @@ part of 'example_swagger.swagger.dart';
 
 Order _$OrderFromJson(Map<String, dynamic> json) {
   return Order(
-    id: json['id'] as int ?? 36,
-    petId: json['petId'] as int ?? 36,
-    quantity: json['quantity'] as int ?? 36,
+    id: json['id'] as int? ?? 36,
+    petId: json['petId'] as int? ?? 36,
+    quantity: json['quantity'] as int? ?? 36,
     shipDateTime: json['shipDateTime'] == null
         ? null
         : DateTime.parse(json['shipDateTime'] as String),
     shipDate: json['shipDate'] == null
         ? null
         : DateTime.parse(json['shipDate'] as String),
-    status: orderStatusFromJson(json['status'] as String),
-    complete: json['complete'] as bool ?? false,
+    status: orderStatusFromJson(json['status'] as String?),
+    complete: json['complete'] as bool? ?? false,
   );
 }
 
@@ -43,14 +43,14 @@ Map<String, dynamic> _$OrderToJson(Order instance) {
 
 OrderWithDash _$OrderWithDashFromJson(Map<String, dynamic> json) {
   return OrderWithDash(
-    id: json['id'] as int ?? 36,
-    petId: json['petId'] as int ?? 36,
-    quantity: json['quantity'] as int ?? 36,
+    id: json['id'] as int? ?? 36,
+    petId: json['petId'] as int? ?? 36,
+    quantity: json['quantity'] as int? ?? 36,
     shipDate: json['shipDate'] == null
         ? null
         : DateTime.parse(json['shipDate'] as String),
-    status: orderWithDashStatusFromJson(json['status'] as String),
-    complete: json['complete'] as bool ?? false,
+    status: orderWithDashStatusFromJson(json['status'] as String?),
+    complete: json['complete'] as bool? ?? false,
   );
 }
 
@@ -74,8 +74,8 @@ Map<String, dynamic> _$OrderWithDashToJson(OrderWithDash instance) {
 
 Category _$CategoryFromJson(Map<String, dynamic> json) {
   return Category(
-    id: json['id'] as int ?? 36,
-    name: json['name'] as String,
+    id: json['id'] as int? ?? 36,
+    name: json['name'] as String?,
   );
 }
 
@@ -95,14 +95,14 @@ Map<String, dynamic> _$CategoryToJson(Category instance) {
 
 User _$UserFromJson(Map<String, dynamic> json) {
   return User(
-    id: json['id'] as int ?? 36,
-    username: json['username'] as String,
-    firstName: json['firstName'] as String,
-    lastName: json['lastName'] as String,
-    email: json['email'] as String,
-    password: json['password'] as String,
-    phone: json['phone'] as String,
-    userStatus: json['userStatus'] as int ?? 36,
+    id: json['id'] as int? ?? 36,
+    username: json['username'] as String?,
+    firstName: json['firstName'] as String?,
+    lastName: json['lastName'] as String?,
+    email: json['email'] as String?,
+    password: json['password'] as String?,
+    phone: json['phone'] as String?,
+    userStatus: json['userStatus'] as int? ?? 36,
   );
 }
 
@@ -128,8 +128,8 @@ Map<String, dynamic> _$UserToJson(User instance) {
 
 Tag _$TagFromJson(Map<String, dynamic> json) {
   return Tag(
-    id: json['id'] as int ?? 36,
-    name: json['name'] as String,
+    id: json['id'] as int? ?? 36,
+    name: json['name'] as String?,
   );
 }
 
@@ -149,19 +149,21 @@ Map<String, dynamic> _$TagToJson(Tag instance) {
 
 Pet _$PetFromJson(Map<String, dynamic> json) {
   return Pet(
-    id: json['id'] as int ?? 36,
+    id: json['id'] as int? ?? 36,
     category: json['category'] == null
         ? null
         : Category.fromJson(json['category'] as Map<String, dynamic>),
-    name: json['name'] as String,
-    photoUrls:
-        (json['photoUrls'] as List)?.map((e) => e as String)?.toList() ?? [],
-    tags: (json['tags'] as List)
+    name: json['name'] as String?,
+    photoUrls: (json['photoUrls'] as List<dynamic>?)
+            ?.map((e) => e as String?)
+            .toList() ??
+        [],
+    tags: (json['tags'] as List<dynamic>?)
             ?.map((e) =>
                 e == null ? null : Tag.fromJson(e as Map<String, dynamic>))
-            ?.toList() ??
+            .toList() ??
         [],
-    status: petStatusFromJson(json['status'] as String),
+    status: petStatusFromJson(json['status'] as String?),
   );
 }
 
@@ -178,16 +180,16 @@ Map<String, dynamic> _$PetToJson(Pet instance) {
   writeNotNull('category', instance.category?.toJson());
   writeNotNull('name', instance.name);
   writeNotNull('photoUrls', instance.photoUrls);
-  writeNotNull('tags', instance.tags?.map((e) => e?.toJson())?.toList());
+  writeNotNull('tags', instance.tags?.map((e) => e?.toJson()).toList());
   writeNotNull('status', petStatusToJson(instance.status));
   return val;
 }
 
 ApiResponse _$ApiResponseFromJson(Map<String, dynamic> json) {
   return ApiResponse(
-    code: json['code'] as int ?? 36,
-    type: json['type'] as String,
-    message: json['message'] as String,
+    code: json['code'] as int? ?? 36,
+    type: json['type'] as String?,
+    message: json['message'] as String?,
   );
 }
 
