@@ -268,7 +268,7 @@ abstract class SwaggerModelsGenerator {
 
     final jsonKeyContent =
         "@JsonKey(name: '$propertyName'$includeIfNullString$dateToJsonValue$unknownEnumValue)\n";
-    return '\t$jsonKeyContent\tfinal $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
+    return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
   String generateUnknownEnumValue(List<String> allEnumNames,
@@ -342,7 +342,7 @@ abstract class SwaggerModelsGenerator {
     final jsonKeyContent =
         "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue$dateToJsonValue)\n";
 
-    return '\t$jsonKeyContent\tfinal $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
+    return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
   String generatePropertyContentByRef(
@@ -373,12 +373,16 @@ abstract class SwaggerModelsGenerator {
     final unknownEnumValue = generateUnknownEnumValue(
         allEnumNames, allEnumListNames, typeName, false);
 
+    if (allEnumListNames.contains(typeName)) {
+      typeName = 'List<$typeName?>';
+    }
+
     final includeIfNullString = generateIncludeIfNullString(options);
 
     final jsonKeyContent =
         "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue)\n";
 
-    return '\t$jsonKeyContent\tfinal $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
+    return '\t$jsonKeyContent\tfinal $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};';
   }
 
   String generateEnumPropertyContent(
@@ -400,7 +404,7 @@ abstract class SwaggerModelsGenerator {
 
     return '''
   @JsonKey($unknownEnumValue$includeIfNullString)
-  final ${className.capitalize + key.capitalize} ${SwaggerModelsGenerator.generateFieldName(key)};''';
+  final ${className.capitalize + key.capitalize}? ${SwaggerModelsGenerator.generateFieldName(key)};''';
   }
 
   String generateListPropertyContent(
@@ -452,7 +456,7 @@ abstract class SwaggerModelsGenerator {
           "@JsonKey(name: '$propertyKey'$includeIfNullString$unknownEnumValue)\n";
     }
 
-    return '''  $jsonKeyContent  final List<$typeName> ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
+    return '''  $jsonKeyContent  final List<$typeName?>? ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
   }
 
   String generateGeneralPropertyContent(
@@ -498,7 +502,7 @@ abstract class SwaggerModelsGenerator {
       jsonKeyContent += ')\n';
     }
 
-    return '''  $jsonKeyContent  final $typeName ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
+    return '''  $jsonKeyContent  final $typeName? ${SwaggerModelsGenerator.generateFieldName(propertyName)};''';
   }
 
   String generatePropertyContentByType(
@@ -676,28 +680,32 @@ abstract class SwaggerModelsGenerator {
     final toLowerCaseString = !enumsCaseSensitive ? '.toLowerCase()' : '';
 
     return '''
-String ${neededName.camelCase}ToJson(enums.$neededName ${neededName.camelCase}) {
+String? ${neededName.camelCase}ToJson(enums.$neededName? ${neededName.camelCase}) {
   return enums.\$${neededName}Map[${neededName.camelCase}];
 }
 
-enums.$neededName ${neededName.camelCase}FromJson(String ${neededName.camelCase}) {
+enums.$neededName ${neededName.camelCase}FromJson(String? ${neededName.camelCase}) {
 
   if(${neededName.camelCase} == null)
   {
     return enums.$neededName.swaggerGeneratedUnknown;
   }
 
+  if (!enums.\$${neededName}Map.entries.contains(${neededName.camelCase})) {
+    return enums.$neededName.swaggerGeneratedUnknown;
+  }
+
   return enums.\$${neededName}Map.entries
-      .firstWhere((element) => element.value$toLowerCaseString == ${neededName.camelCase}$toLowerCaseString, orElse: () => null)
-      ?.key ?? enums.$neededName.swaggerGeneratedUnknown;
+      .firstWhere((element) => element.value$toLowerCaseString == ${neededName.camelCase}$toLowerCaseString)
+      .key;
 }
 
-List<String> ${neededName.camelCase}ListToJson(
-    List<enums.$neededName> ${neededName.camelCase}) {
+List<String?>? ${neededName.camelCase}ListToJson(
+    List<enums.$neededName?>? ${neededName.camelCase}) {
 
   if(${neededName.camelCase} == null)
   {
-    return null;
+    return [];
   }
 
   return ${neededName.camelCase}
@@ -705,8 +713,8 @@ List<String> ${neededName.camelCase}ListToJson(
       .toList();
 }
 
-List<enums.$neededName> ${neededName.camelCase}ListFromJson(
-    List ${neededName.camelCase}) {
+List<enums.$neededName?>? ${neededName.camelCase}ListFromJson(
+    List? ${neededName.camelCase}) {
 
   if(${neededName.camelCase} == null)
   {
